Endpoints for discharges, placements, referrals
