+++ app/controllers/in_care_controller.rb
@@ -1,0 +1,14 @@
+require 'enrollments_helper'
+
+class InCareController < GeneralEnrollmentsController
+
+  def correct_params
+    params["in_care"] = true
+  end
+
+  def count
+    self.query
+    render json: {:in_care => @enrollments.count}
+  end
+
+end
