+++ app/controllers/placements_controller.rb
@@ -1,0 +1,9 @@
+require 'enrollments_helper'
+
+class PlacementsController < GeneralEnrollmentsController
+
+  def correct_params
+    params["date_orr_approved"] = params.delete "on"
+  end
+
+end
