+++ app/controllers/referrals_controller.rb
@@ -1,0 +1,9 @@
+require 'enrollments_helper'
+
+class ReferralsController < GeneralEnrollmentsController
+
+  def correct_params
+    params["referral_date"] = params.delete "on"
+  end
+
+end
